Removed the pseudobatch_transform_by_group() funciton.
This was done because we believe that it is better to teach how to use the native pandas, because it give user more flexibility and reduce maintainance.

The tutorial 6 - Real data... was adapted to this change. However, the article_figures notebook wasn't because it is adapted in another incoming change from branch improve_figure.