Dockerfile produce functioning container for both Julia and CmdStan
The Dockerfile now produces a separate conda environment were all packages are installed. Furthermore, the compiled stan model is deleted, thus they need to be recompiled inside the container. This was required for the model to work.

The Python dependencies are installed to match the exact versions of the environment used to produce the results from the article. These depencies are decribed in the article/requirements.txt.

The error propagation notebook now starts with a block that detects if it is run inside the docker container, if so it set the cmdstan path.

The Julia setup has been shrink significantly, now it no longer updates the packages, but uses the versions specified in the project.toml file. Unfortunately, the Julia packages is not installed during image build, but this is instead done in the julia_run_all.jl file.